Initial Project Setup
Set up the initial project structure using "npx create-react-app amazon-clone"

'npx' is package runner tool that comes with npm. It is used to execute binaries from npm packages.

'create-react-app' is a command-line utility that sets up a new React project with a recommended project structure and initial configurations. It simplifies the process of creating a new React application

The full command fetches the latest version of 'create-react-app' from the npm registry, uses it to create a new React project in a directory named 'amazon-clone', and sets up the necessary files and dependencies for a basic React application.

Also added a firebase.js file with code for setting up the Firebase DB.